% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_bcadata.R
\name{compare_bcadata}
\alias{compare_bcadata}
\title{Reconstruct BCA data from raw sources: FRED and TFP}
\usage{
compare_bcadata(bcadata)
}
\arguments{
\item{bcadata}{data.frame from pull_bcadata()}
}
\value{
A ggplot
}
\description{
Reconstruct BCA data from raw sources: FRED and TFP
}
\examples{
\dontrun{
data <- compare_bcadata(pull_bcadata("XXXXXXXXXXX"))
}
}
