% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_var_results.R
\docType{data}
\name{original_var_results}
\alias{original_var_results}
\title{Selected Original Business Cycle Anatomy VAR Impulse Response}
\format{
\subsection{\code{original_bcadata}}{

A data frame with 1600 rows and 6 columns:
\describe{
\item{horizon}{Impulse response horizon, in quarters}
\item{median}{Median response of Bayesian iterations}
\item{pctl_84}{Upper confidence band for 68\%}
\item{pctl_84}{Lower confidence band for 68\%}
\item{variable}{One of ten macro variables responding to the shock}
\item{model}{VAR model trained as Bayesian or classical,
targetting frequency domain (fd) or time domain (td)}
}
}
}
\source{
\url{https://www.openicpsr.org/openicpsr/project/118082/version/V1/view}
}
\usage{
original_var_results
}
\description{
Pulled from the replication files and transformed into a tibble.
Impulse response to the "Main Business Cycle" shock of
ten U.S. macro time series, for a horizon of 40 quarters.
There are four different VAR models included.
}
\keyword{datasets}
