% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tfp_data.R
\name{pull_tfp_data}
\alias{pull_tfp_data}
\title{Pull TFP data from San Franciso Fed website}
\usage{
pull_tfp_data(replicate = FALSE)
}
\arguments{
\item{replicate}{Boolean, use original paper vintage}
}
\value{
A tibble/data.frame
}
\description{
Pull TFP data from San Franciso Fed website
}
