% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Merge FRED, TFP data and construct correct series}
\usage{
transform_data(fred_data, tfp_data)
}
\arguments{
\item{fred_data}{Data.frame from pull_fred_data}

\item{tfp_data}{Data.frame from pull_tfp_data}
}
\value{
A tibble/data.frame
}
\description{
Merge FRED, TFP data and construct correct series
}
