% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_bcadata.R
\name{pull_bcadata}
\alias{pull_bcadata}
\title{Reconstruct BCA data from raw sources: FRED and TFP}
\usage{
pull_bcadata(fred_api_key)
}
\arguments{
\item{fred_api_key}{API key from FRED}
}
\value{
A tibble/data.frame
}
\description{
Reconstruct BCA data from raw sources: FRED and TFP
}
\examples{
\dontrun{
data <- pull_bcadata("XXXXXXXXXXX")
}
}
